+++ Source/WebCore/platform/xr/openxr/PlatformXROpenXR.h
@@ -20,14 +20,16 @@
 #pragma once
 
 #if ENABLE(WEBXR) && USE(OPENXR)
+
+#include "OpenXRUtils.h"
 #include "PlatformXR.h"
 
 #include <wtf/HashMap.h>
-
-#include <openxr/openxr.h>
 #include <wtf/WorkQueue.h>
 
 namespace PlatformXR {
+
+class OpenXRExtensions;
 
 // https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#system
 // A system represents a collection of related devices in the runtime, often made up of several individual
@@ -43,46 +45,41 @@ namespace PlatformXR {
 // the XRSystem is basically the entry point for the WebXR API available via the Navigator object.
 class OpenXRDevice final : public Device {
 public:
-    OpenXRDevice(XrSystemId, XrInstance, WorkQueue&, CompletionHandler<void()>&&);
-    XrSystemId xrSystemId() const { return m_systemId; }
+    OpenXRDevice(XrInstance, XrSystemId, WorkQueue&, const OpenXRExtensions&, CompletionHandler<void()>&&);
 
 private:
-    void collectSupportedSessionModes();
-    void collectConfigurationViews();
-
-    ListOfEnabledFeatures enumerateReferenceSpaces(XrSession&) const;
-    XrSpace createReferenceSpace(XrReferenceSpaceType);
-
+    // PlatformXR::Device
     WebCore::IntSize recommendedResolution(SessionMode) final;
-
     void initializeTrackingAndRendering(SessionMode) final;
     void shutDownTrackingAndRendering() final;
     void initializeReferenceSpace(PlatformXR::ReferenceSpaceType) final;
     bool supportsSessionShutdownNotification() const final { return true; }
-    void waitUntilStopping();
+    void requestFrame(RequestFrameCallback&&) final;
+    Vector<ViewData> views(SessionMode) const final;
 
+    // Custom methods
+    ListOfEnabledFeatures enumerateReferenceSpaces(XrSession) const;
+    void collectSupportedSessionModes();
+    void collectConfigurationViews();
+    XrSpace createReferenceSpace(XrReferenceSpaceType);
     void pollEvents();
     XrResult beginSession();
     void endSession();
     void resetSession();
     void handleSessionStateChange();
+    void waitUntilStopping();
 
-    void requestFrame(RequestFrameCallback&&) final;
-
-    Vector<ViewData> views(SessionMode) const final;
+    XrInstance m_instance;
+    XrSystemId m_systemId;
+    WorkQueue& m_queue;
+    const OpenXRExtensions& m_extensions;
+    XrSession m_session { XR_NULL_HANDLE };
+    XrSessionState m_sessionState { XR_SESSION_STATE_UNKNOWN };
 
     using ViewConfigurationPropertiesMap = HashMap<XrViewConfigurationType, XrViewConfigurationProperties, IntHash<XrViewConfigurationType>, WTF::StrongEnumHashTraits<XrViewConfigurationType>>;
     ViewConfigurationPropertiesMap m_viewConfigurationProperties;
     using ViewConfigurationViewsMap = HashMap<XrViewConfigurationType, Vector<XrViewConfigurationView>, IntHash<XrViewConfigurationType>, WTF::StrongEnumHashTraits<XrViewConfigurationType>>;
     ViewConfigurationViewsMap m_configurationViews;
-
-    XrSystemId m_systemId;
-    XrInstance m_instance;
-    XrSession m_session { XR_NULL_HANDLE };
-    XrSessionState m_sessionState { XR_SESSION_STATE_UNKNOWN };
-
-    WorkQueue& m_queue;
-
     XrViewConfigurationType m_currentViewConfigurationType;
     XrSpace m_localSpace { XR_NULL_HANDLE };
     XrSpace m_viewSpace { XR_NULL_HANDLE };
